# chatgpt-app/openapi.yaml
openapi: "3.1.0"
info:
  title: Course Companion FTE API
  description: >
    Zero-Backend-LLM API for the AI Agent Development course by Panaversity.
    All intelligent reasoning is handled by the ChatGPT App (frontend LLM).
    This backend is 100% deterministic — no LLM calls.
  version: "1.0"
servers:
  - url: https://your-backend.railway.app
    description: Production server (update after Railway deployment)

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/content/chapters/{chapter_id}:
    get:
      operationId: get_chapter_content
      summary: Get full chapter content
      description: >
        Returns the full markdown content of a chapter.
        Checks freemium access gate first. If the user does not have the
        required tier, returns 403. After serving, automatically updates
        the user's last_read progress.
      parameters:
        - name: chapter_id
          in: path
          required: true
          description: Chapter identifier (e.g. chapter-01)
          schema:
            type: string
            example: chapter-01
        - name: X-User-ID
          in: header
          required: false
          description: The user's unique identifier for access control and progress tracking
          schema:
            type: string
            example: chatgpt_user_abc123
      responses:
        "200":
          description: Full chapter markdown content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterContent"
        "403":
          description: User does not have the required tier to access this chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessDeniedError"
        "404":
          description: Chapter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "503":
          description: Content not available (R2 not yet seeded)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
  /api/v1/content/chapters/{chapter_id}/summary:
    get:
      operationId: get_chapter_summary
      summary: Get chapter preview (first 200 words)
      description: >
        Returns the first 200 words of a chapter as a preview.
        Available to all tiers — use this to show locked chapter previews
        when the user does not have premium access.
      parameters:
        - name: chapter_id
          in: path
          required: true
          description: Chapter identifier (e.g. chapter-04)
          schema:
            type: string
            example: chapter-04
      responses:
        "200":
          description: Chapter preview (first 200 words)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterSummary"
        "404":
          description: Chapter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  # ── 6.2 Navigation APIs ──────────────────────────────────────────────────────
  /api/v1/navigation/chapters/{chapter_id}/next:
    get:
      operationId: get_next_chapter
      summary: Get next chapter metadata
      description: >
        Returns metadata for the chapter that follows chapter_id
        in the course sequence (based on order_index). Does not return
        full content — use get_chapter_content for that.
      parameters:
        - name: chapter_id
          in: path
          required: true
          description: Current chapter identifier
          schema:
            type: string
            example: chapter-01
      responses:
        "200":
          description: Metadata for the next chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterNavItem"
        "404":
          description: No next chapter (this is the last chapter)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
  /api/v1/navigation/chapters/{chapter_id}/prev:
    get:
      operationId: get_prev_chapter
      summary: Get previous chapter metadata
      description: >
        Returns metadata for the chapter that precedes chapter_id
        in the course sequence. Returns 404 if this is the first chapter.
      parameters:
        - name: chapter_id
          in: path
          required: true
          description: Current chapter identifier
          schema:
            type: string
            example: chapter-03
      responses:
        "200":
          description: Metadata for the previous chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterNavItem"
        "404":
          description: No previous chapter (this is the first chapter)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
...
  /api/v1/access/check/{user_id}/{chapter_id}:
    get:
      operationId: check_access
      summary: Check chapter access for a user
      description: >
        Returns whether a user has access to a specific chapter based on
        their subscription tier. Logic: free chapters are always accessible;
        premium chapters require premium or higher tier; pro requires pro tier.
        Returns 'free' tier if the user is not yet registered.
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
            example: chatgpt_user_abc123
        - name: chapter_id
          in: path
          required: true
          description: Chapter identifier
          schema:
            type: string
            example: chapter-04
      responses:
        "200":
          description: Access check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessCheckResponse"
        "404":
          description: Chapter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
  /api/v1/access/users:
    post:
      operationId: create_or_get_user
      summary: Create or get user
      description: >
        Creates a new user or returns the existing user (upsert by user_id).
        Call this on first interaction to register the ChatGPT user in the system.
        Returns the user object including their current tier.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: User object (created or existing)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
  /api/v1/access/users/{user_id}/tier:
    get:
      operationId: get_user_tier
      summary: Get user subscription tier
      description: >
        Returns the subscription tier for a user_id.
        Returns 'free' if the user is not registered in the system.
        Possible values: free | premium | pro | team
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
            example: chatgpt_user_abc123
      responses:
        "200":
          description: User tier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TierResponse"

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS / SCHEMAS
# ─────────────────────────────────────────────────────────────────────────────
components:
  schemas:
...
    ChapterContent:
      type: object
      properties:
        chapter_id:
          type: string
          example: chapter-01
        title:
          type: string
        tier_required:
          type: string
          enum: [free, premium, pro]
        order_index:
          type: integer
        estimated_minutes:
          type: integer
          nullable: true
        content:
          type: string
          description: Full markdown text of the chapter
      required: [chapter_id, title, tier_required, order_index, content]
    ChapterSummary:
      type: object
      properties:
        chapter_id:
          type: string
        title:
          type: string
        tier_required:
          type: string
          enum: [free, premium, pro]
        preview:
          type: string
          description: First 200 words of the chapter
      required: [chapter_id, title, tier_required, preview]

    # ── Navigation ────────────────────────────────────────────────────────────
    ChapterNavItem:
      type: object
      properties:
        chapter_id:
          type: string
          example: chapter-02
        title:
          type: string
          example: Claude Agent SDK Deep Dive
        tier_required:
          type: string
          enum: [free, premium, pro]
        order_index:
          type: integer
          example: 2
      required: [chapter_id, title, tier_required, order_index]
...
    AccessCheckResponse:
      type: object
      properties:
        has_access:
          type: boolean
          example: true
        tier_required:
          type: string
          enum: [free, premium, pro]
          example: free
        user_tier:
          type: string
          enum: [free, premium, pro, team]
          example: free
        chapter_id:
          type: string
          example: chapter-01
        user_id:
          type: string
          example: chatgpt_user_abc123
      required: [has_access, tier_required, user_tier, chapter_id, user_id]
    CreateUserRequest:
      type: object
      properties:
        user_id:
          type: string
          example: chatgpt_user_abc123
        email:
          type: string
          format: email
          nullable: true
          example: student@example.com
      required: [user_id]
    UserOut:
      type: object
      properties:
        user_id:
          type: string
          example: chatgpt_user_abc123
        email:
          type: string
          nullable: true
        tier:
          type: string
          enum: [free, premium, pro, team]
          example: free
        created_at:
          type: string
          nullable: true
          example: "2026-04-11T10:00:00"
      required: [user_id, tier]
    TierResponse:
      type: object
      properties:
        user_id:
          type: string
          example: chatgpt_user_abc123
        tier:
          type: string
          enum: [free, premium, pro, team]
          example: free
      required: [user_id, tier]

    # ── Errors ────────────────────────────────────────────────────────────────
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
          example: Chapter 'chapter-99' not found
      required: [detail]
    AccessDeniedError:
      type: object
      properties:
        detail:
          type: object
          properties:
            error:
              type: string
              example: premium_required
            upgrade_message:
              type: string
              example: This chapter requires Premium tier. Your current tier is free.
            tier_required:
              type: string
              enum: [free, premium, pro]
              example: premium
            user_tier:
              type: string
              enum: [free, premium, pro, team]
              example: free
          required: [error, upgrade_message, tier_required, user_tier]
      required: [detail]